ppo_basic mostly works. Should be the preferred one. It has a way to load the last saved run for next run. Limit on the number of steps can also be mentioned.
Once trained, it can be run by loading the model and running it using gym,a nd if it works well can be run by transforming state space on the server.

If logging is needed, might need to setup wandb, and gpu access for pytorch needs to be setup. Will also add comments and a new file for creating the env on conda