still the same. but its learned some kind of defensive policy. It does not collect coins, but is able to move away from a longer snake.
There is some self killing present, possibly due to sparse rewards.

There is also no cooperation between the agents. That has to be worked on. Maybe partial penalties if enemy kills your teammate for that step. It might help improve performance.

Have to check out. I am going to train this further because at some point I believe it will learn to collect food, at which point it could learn to fully win games. And once that happens it should be able to cosntantly improve itself